Display progress dialog when fetching secrets
Fetching secrets might add a delay before the process is
executed. This might be confusing without any UI change
(only the button was disabled).

At the moment it's not possible to cancel the fetching
of secrets. I tried to pass ProgressIndicator to the
http utils, but the cancellation didn't work there.