Fix handling of TLS certificate errors
The ktor implementation has changed how TLS certificate errors
are handled. To fix, we use the IDE's trust manager so that
the same handling is used as in other parts of the IDE.
As a workaround for showing the dialog multiple times, the first
fetch of a secret from a batch is done serially.